Renamed flag to goal
